[LCR-142] refactor: 결제 서비스 FeignClient를 infra 계층으로 이동하여 도메인 간 중복 제거
